- handle page reloading - better components render - avoid double rendering - data from localstorage with type
