+++ frontend/src/components/Layout.tsx
@@ -11,6 +11,7 @@ import { useAuth } from '@/context/AuthProvider';
 import { LoginForm } from '@/components/LoginForm';
 import { toast } from 'sonner';
 import { redirect } from 'react-router-dom';
+import { ConfigProvider } from '@/context/ConfigContext';
 
 export default function Layout() {
   const { theme } = useTheme();
@@ -39,56 +40,58 @@ export default function Layout() {
   }, []);
 
   return (
-    <ThemeProvider defaultTheme="dark" storageKey="teleskopio-ui-theme">
-      {AuthDisabled ? (
-        <div className="group/sidebar-wrapper has-data-[variant=inset]:bg-sidebar flex min-h-svh w-full">
-          <SidebarProvider>
-            <AppSidebar />
-          </SidebarProvider>
-          <main className="bg-background flex w-full flex-col h-screen">
-            <Outlet />
-          </main>
-        </div>
-      ) : (
-        <div className="group/sidebar-wrapper has-data-[variant=inset]:bg-sidebar flex min-h-svh w-full">
-          {isAuthenticated ? (
+    <ConfigProvider>
+      <ThemeProvider defaultTheme="dark" storageKey="teleskopio-ui-theme">
+        {AuthDisabled ? (
+          <div className="group/sidebar-wrapper has-data-[variant=inset]:bg-sidebar flex min-h-svh w-full">
             <SidebarProvider>
               <AppSidebar />
             </SidebarProvider>
-          ) : (
-            <></>
-          )}
-          <main className="bg-background flex w-full flex-col h-screen">
-            {isAuthenticated ? (
+            <main className="bg-background flex w-full flex-col h-screen">
               <Outlet />
+            </main>
+          </div>
+        ) : (
+          <div className="group/sidebar-wrapper has-data-[variant=inset]:bg-sidebar flex min-h-svh w-full">
+            {isAuthenticated ? (
+              <SidebarProvider>
+                <AppSidebar />
+              </SidebarProvider>
             ) : (
-              <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
-                <div className="w-full max-w-sm">
-                  <LoginForm login={handleLogin} />
-                </div>
-              </div>
+              <></>
             )}
-          </main>
-        </div>
-      )}
-      <Toaster
-        icons={{
-          success: <ThumbsUp color="green" size={18} />,
-          info: <Info size={18} />,
-          warning: <CircleAlert color="orange" size={18} />,
-          error: <OctagonMinus color="red" size={18} />,
-          loading: <Loader size={18} />,
-        }}
-        theme={theme}
-        visibleToasts={3}
-        toastOptions={{
-          className: '!text-xs',
-          style: {
-            fontFamily: 'var(--app-font)',
-          },
-        }}
-        position="bottom-right"
-      />
-    </ThemeProvider>
+            <main className="bg-background flex w-full flex-col h-screen">
+              {isAuthenticated ? (
+                <Outlet />
+              ) : (
+                <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
+                  <div className="w-full max-w-sm">
+                    <LoginForm login={handleLogin} />
+                  </div>
+                </div>
+              )}
+            </main>
+          </div>
+        )}
+        <Toaster
+          icons={{
+            success: <ThumbsUp color="green" size={18} />,
+            info: <Info size={18} />,
+            warning: <CircleAlert color="orange" size={18} />,
+            error: <OctagonMinus color="red" size={18} />,
+            loading: <Loader size={18} />,
+          }}
+          theme={theme}
+          visibleToasts={3}
+          toastOptions={{
+            className: '!text-xs',
+            style: {
+              fontFamily: 'var(--app-font)',
+            },
+          }}
+          position="bottom-right"
+        />
+      </ThemeProvider>
+    </ConfigProvider>
   );
 }
