+++ frontend/src/components/resources/PaginatedTable.tsx
@@ -13,14 +13,15 @@ interface PaginatedTableProps<T> {
   kind: string;
   group: string;
   getPage: (args: {
-    apiResource: ApiResource | undefined;
+    server: string | undefined;
     limit: number;
     continueToken?: string;
   }) => Promise<[T[], string | null, string]>;
-  subscribeEvents: (rv: string, apiResource: ApiResource | undefined) => Promise<void>;
+  subscribeEvents: (rv: string) => Promise<void>;
   state: () => Map<string, T>;
   setState: (updater: (prev: Map<string, T>) => Map<string, T>) => void;
   extractKey: (item: T) => string;
+  apiResource: ApiResource | undefined;
   columns: any;
   namespaced?: boolean;
   withNsSelector?: boolean;
@@ -35,6 +36,7 @@ export function PaginatedTable<T>({
   group,
   subscribeEvents,
   state,
+  apiResource,
   setState,
   extractKey,
   columns,
@@ -48,25 +50,15 @@ export function PaginatedTable<T>({
   const loaderRef = useRef<HTMLDivElement | null>(null);
   const selectedNamespace = useSelectedNamespacesState();
   const [searchQuery, setSearchQuery] = useState('');
-  const { serverInfo } = useConfig();
-
-  const getApiResource = ({
-    kind,
-    group,
-  }: {
-    kind: string;
-    group: string;
-  }): ApiResource | undefined => {
-    return serverInfo?.apiResources.find((r: ApiResource) => r.kind === kind && r.group === group);
-  };
+  const { serverInfo, isLoading } = useConfig();
 
   const loadPage = async () => {
     if (loading) return;
     setLoading(true);
     try {
-      const apiResource = getApiResource({ kind, group });
+      if (serverInfo?.server === '') return;
       const [items, next, rv] = await getPage({
-        apiResource: apiResource,
+        server: serverInfo?.server,
         limit: 50,
         continueToken: nextToken ?? undefined,
       });
@@ -77,7 +69,7 @@ export function PaginatedTable<T>({
         });
         return newMap;
       });
-      await subscribeEvents(rv, apiResource);
+      await subscribeEvents(rv);
       setNextToken(next);
     } catch (e: any) {
       console.error('Error loading page:', e);
@@ -91,7 +83,7 @@ export function PaginatedTable<T>({
 
   useEffect(() => {
     loadPage();
-  }, []);
+  }, [isLoading]);
 
   useEffect(() => {
     if (!loaderRef.current || !nextToken) return;
@@ -130,11 +122,7 @@ export function PaginatedTable<T>({
         .includes(searchQuery.toLowerCase());
     })
     .filter(
-      (x: any) =>
-        !getApiResource({ kind, group })?.namespaced ||
-        !ns ||
-        ns === 'all' ||
-        x.metadata.namespace === ns,
+      (x: any) => !apiResource?.namespaced || !ns || ns === 'all' || x.metadata.namespace === ns,
     );
   const showInitialLoader = loading && data.length === 0;
   return (
@@ -148,7 +136,7 @@ export function PaginatedTable<T>({
       <div className="flex-1 overflow-y-auto">
         <DataTable
           kind={kind}
-          apiResource={getApiResource({ kind, group })}
+          apiResource={apiResource}
           doubleClickDisabled={doubleClickDisabled}
           noResult={data.length === 0}
           columns={columns}
